+++ test/e2e/helpers.js
@@ -2,8 +2,8 @@ const path = require('path');
 const BigNumber = require('bignumber.js');
 const mockttp = require('mockttp');
 const createStaticServer = require('../../development/create-static-server');
-const { setupMocking } = require('../../development/mock-e2e');
 const enLocaleMessages = require('../../app/_locales/en/messages.json');
+const { setupMocking } = require('./mock-e2e');
 const Ganache = require('./ganache');
 const FixtureServer = require('./fixture-server');
 const { buildWebDriver } = require('./webdriver');
